Add "configtest" CLI command.
This adds a "configtest" command-line argument that sets up an instance
of Headscale and exits, causing the configuration file and ACLs to
be processed.

This solves a problem I've been having where the only way I know to
test the ACLs syntax is to restart headscale, and if it fails to start
I need to login via another system and correct the problem, then restart
headscale.

This is the first time I've touched go, so if you ask too much of me in
review I'm likely to be out of my element quickly.  :-)