Add unix_socket directive to the "likely wanted changes" area
The default config places the control socket in /var/lib/headscale/headscale.sock, causing an error. As noted for the other files this area is not writable in the docker image. I also fixed a double space while I was at it.

I'm not sure if protocol is to also create an issue and link it to this, but please let me know and I'll do that if it's needed.