Add github stale action
We have a lot of issues that has gone stale, and to help us
prioritise which ones to pick up, we are adding this stale
action to close out issues that are no longer active.

The bot is configured as this:
- If an issue sees no activity for 6 months (180 days) it is marked as stale
- If an issue is not confirmed still being an issue within 14 days, it is closed.

This is of course not a perfect solution, but it will hopefully
help us prioritise active issues with people who can help us debug.

We might tighten the activity requirement in the future.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>