Rework state and online system
This PR further improves the state management system and tries to
make sure that we get all nodes in sync continously.

This is greatly enabled by a previous PR dropping support for older
clients that allowed us to use a Patch field only sending small diffs
for client updates.

It also reworks how the HA subnet router is handled and it should
be a bit easier to follow now.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>