add versioned migrations
Up until this commit, headscale has had one massive migration func that
runs every time the server starts. This means that every time we have
migrated anything, we have added stuff to it and not really taken into
account how the database of clients look like. This means that sometimes
the new stuff has had to be added in the middle, as it would break things
further down.

It is kind of a miracle that it has worked so far with no permament damage,
particularly considering some of our larger migrations like Machine -> Node
or Namespace -> User.

This commit addresses this by adding gormigrate, which is a simple versioned
migration system. As we have to maintain our messy past, the first migration
is the large, currently working, but hard to change function. Subsequent changes
should be added with new IDs after this.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>