Fix and test node expiry
This commit makes sure node expiry is propagated to
the nodes immediatly when a node expiry even occurs
and verifies this with integration tests.

In addition, it removes old logic for filtering out
expired nodes from the netmap, this is not how Tailscale
SaaS does it, and we should follow the same approach.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>