introduce dedicated ip allocator
This new allocator keeps the used space in memory and
loads the database only on startup. If it allocates but
something errors before it gets saved to the database,
then thats fine because it will be cleared on next restart.

This should resolve races between ip allocation with postgres
and as a side effect should make it easier to implement random
allocation instead of serial if we so desire.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>