remove mapresponse map
this commit removes the mapresponse map as it complicated
and caused more problems than it solved.

instead, notifier is now changed to always use the last
connected nodes channel and to ensure it is not remove until
that client is disconnected.

as part of this, offline/route change is a bit more concervative
and only get sent if the channel was removed from the notifier
and not if the clients longpolling session ended (there might be
more than one at the time, particularly if the client is connecting.)

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>