feat: implements apis for managing headscale policy
This commit introduces APIs for managing the headscale policy. Until
now, the support was limited to file based policy and users could only
update the file and reload headscale in order to apply the changes in
the policy. With the new APIs, the users of headscale and now manage the
policy programatically and will not require to restart headscale.

The commit also implements the CLI commands to get and set the policy.

BREAKING CHANGE: headscale no longer supports YAML policy files and the
only supported format is HuJSON.