Only load needed part of configuration
Currently all configuration was loaded and validated regardless of what
the CLI was doing, if you did a `nodes ls` or `serve` the whole thing would
be loaded and validated which caused a machine that were doing only CLI
stuff to need the full config, or at least a bunch that would throw errors.

This commit cleans up and splits this so that only the subset needed for
running the CLI (unix socket or remote gRPC) is required when running
all commands except for `serve`.

When `serve` is ran, the normal chain with validations and loading of the
whole config is ran.

This means that deprecation warnings for the server config will only be shown
in the server log.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>