all: remove 32 bit support
This commit removes the official 32 bit builds.
As per #2511, they are broken at runtime and we
do not have any good way to ensure that they run over
time as we cannot catch these issue at compile or
test time.

64-bit has been around for long enough that this
change should not be too disruptive.

32-bit users can compile the code themselves, and
we can still be open to accept patches for it, but
not provide builds as they might not work over time.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>