feat: Add `magicdns_peer_aaaa` config option
This change introduces a new boolean config option,
`magicdns_peer_aaaa`.

When enabled, it adds the `magicdns-aaaa` node attribute to devices with
a valid overlay IPv6 address.
This attribute allows MagicDNS to resolve AAAA (IPv6) DNS queries for
other nodes in the tailnet.

This is a new opt-in feature for Tailscale 1.84+ and is implemented as a
global config option, similar to `RandomizeClientPort`, since support
for configuring node attributes via policies is not yet available.