cli: only validate bypass-grpc set policy
When adding the mode to bypass gRPC, a naive policy validation
was added to both the gRPC set, and the bypass mode.

This caused the policy to invalidate in unexpected ways as the gRPC
already handles this validation in the backend.

This commit moves the validation logic into the bypass branch.

Fixes #2825

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>