types: make pre auth key use bcrypt
this commit changes the underlying implementation of
pre auth keys to the same as our API keys. This means that
instead of having a key in the database, it will consist
of a prefix and a bcrypt hash.

This improves the security of the keys as they can not be
read back in the case of a database getting away.

Old keys still work, but all new keys are created in the
new format.

All keys are prefixed with "hskey-auth" to indicate their
purpose.

Signed-off-by: Kristoffer Dalby <kristoffer@tailscale.com>