Separate the concepts of wg-only peers, and their connections to tailnet nodes
This allows us to properly utilize the idea of multiple KnownNodeIDs for
a single WireGuard node. Limitations in the previous types required all
of them to share a single masquerade address which isn't realistic.

A single WireGuard-only node has some properties which are essential to
this peer (e.g. public key, endpoints), and some properties which vary
depending on which tailnet node wants to connect to it (e.g. masquerade
address). This commit adjusts the whole wg-only peer feature to deal
with this logical separation.