tags: separate user and tag owned nodes
This commit changes tags to be something that exists on nodes in addition to users, to being its own thing. It is part
of moving our tags support towards the correct tailscale compatible implementation.

There are probably rough edges in this commit, but the intention is to get it in, and then start fixing bugs from 0.28.0
milestone (long standing tags issue) to discover what works and what doesnt.

Updates #2417
Closes #2619