cli: remove node move command
This commit removes support for moving nodes between users.

The reason for this is that allowing such move is going against
a intended design about identity in Tailscale, where your peer
has an identity based on their machine, and user id. By allowing
the node to be moved, we are breaking this contract, a user might
have a peer, and it cannot be trusting who owns this peer as it
cant change.

This feature should not have been added, and now that Tags are being
implemented, essentially representing service account, allowing better
seggeration anyways, we will remove it.

One of the problems associated with this would be similar to the email
takeover on an OIDC platform.

Signed-off-by: Kristoffer Dalby <kristoffer@dalby.cc>